Do not match newlines in `highlight` output
Without this, `links to Foo` might be included, which results in an
error, when throwing that into a call to `highlight` to restore it.

The proper fix would be to restore it as `highlight link Cursor|Error
FooLink`, but that requires refactoring, since the return value for
`hl_cursor_off`/`turn_off_hl_error` is used as a definition also.

Fixes https://github.com/osyo-manga/vital-over/issues/23
Fixes https://github.com/Lokaltog/vim-easymotion/issues/95