Add Vim-Plug and Vim8 Native installation help
Vim-Plug is one of the most popular Vim plugin managers by stars on Github, and so should be one of the methods to have its usage described in the installation section.
Vim8 has native plugin management, allowing all Vim8 users to use it without installing additional plugins. For this reason a usage method should also be added.